03-apps/discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "1.0.0"
description = "Discord bot for Immich integration"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.1",
    "fastmcp>=2.0.0,<3.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "atproto>=0.0.50",
    "pytumblr>=0.1.0",
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bot"]
